requirements.txt
python-telegram-bot[job-queue]>=21.0,<22.0
openai>=1.30.0
langfuse>=2.0.0
python-dotenv>=1.0.0
